move filterParentEntities to EntityTree
