--- packages/editor/src/functions/filterParentEntities.ts
@@ -1,75 +1,0 @@
-/*
-CPAL-1.0 License
-
-The contents of this file are subject to the Common Public Attribution License
-Version 1.0. (the "License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-https://github.com/EtherealEngine/etherealengine/blob/dev/LICENSE.
-The License is based on the Mozilla Public License Version 1.1, but Sections 14
-and 15 have been added to cover use of software over a computer network and 
-provide for limited attribution for the Original Developer. In addition, 
-Exhibit A has been modified to be consistent with Exhibit B.
-
-Software distributed under the License is distributed on an "AS IS" basis,
-WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the
-specific language governing rights and limitations under the License.
-
-The Original Code is Ethereal Engine.
-
-The Original Developer is the Initial Developer. The Initial Developer of the
-Original Code is the Ethereal Engine team.
-
-All portions of the code written by the Ethereal Engine team are Copyright © 2021-2023 
-Ethereal Engine. All Rights Reserved.
-*/
-
-import { getComponent, hasComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { Entity } from '@etherealengine/ecs/src/Entity'
-import { EntityTreeComponent } from '@etherealengine/spatial/src/transform/components/EntityTree'
-import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
-
-/**
- * Filters the parent entities from the given entity list.
- * In a given entity list, suppose 2 entities has parent child relation (can be any level deep) then this function will
- * filter out the child entity.
- * @param nodeList List of entities to find parents from
- * @param parentNodeList Resulter parent list
- * @param filterUnremovable Whether to filter unremovable entities
- * @param filterUntransformable Whether to filter untransformable entities
- * @returns List of parent entities
- */
-export const filterParentEntities = (
-  rootEntity: Entity,
-  entityList: Entity[],
-  parentEntityList: Entity[] = [],
-  filterUnremovable = true,
-  filterUntransformable = true
-): Entity[] => {
-  parentEntityList.length = 0
-
-  // Recursively find the nodes in the tree with the lowest depth
-  const traverseParentOnly = (entity: Entity) => {
-    if (!entity) return
-
-    const node = getComponent(entity, EntityTreeComponent)
-
-    if (
-      entityList.includes(entity) &&
-      !(filterUnremovable && !node.parentEntity) &&
-      !(filterUntransformable && !hasComponent(entity, TransformComponent))
-    ) {
-      parentEntityList.push(entity)
-      return
-    }
-
-    if (node.children) {
-      for (let i = 0; i < node.children.length; i++) {
-        traverseParentOnly(node.children[i])
-      }
-    }
-  }
-
-  traverseParentOnly(rootEntity)
-
-  return parentEntityList
-}
